CI: Add step to check minimal versions on detached crate (#181)
#### Problem

Although CI checks that we satisfy the minimal versions declared in
Cargo.toml, it doesn't work for dependencies in the workspace, as
noticed in #175. There, solana-keypair should depend on a newer
version of solana-signature, but the normal minimal-versions check
doesn't catch it.

#### Summary of changes

The script is a bit brittle, but it detaches a crate from the workspace
and then runs `cargo-minimal-versions` on it as a pre-publish test. Once
this lands, I'll do a dry-run publish to make sure solana-keypair fails.