program-error!: Improve ToStr ergonomics (#189)
#### Problem

The `ToStr` trait is a bit difficult to use, forcing a lot of redundant
uses of types, when programs will just want to call
`err.to_str::<MyError>()`.

#### Summary of changes

Simplify the trait to not require anything, then add a generic `to_str`
function directly on `ProgramError` that allows us to get the error
message for a particular error type.

Also, clarify the deprecation warning on `DecodeError`.

BREAKING CHANGE: the `ToStr` trait requires fewer bounds