message / transaction: Remove dependency on solana-system-interface (#239)
* message: Remove dependency on solana-system-interface

#### Problem

While debugging changes to the wasm-js build, I noticed that
`solana-transaction` and `solana-message` were pulling in
`solana-system-interface` in a lot of situations, even though there's a
lot of code in place to make sure that doesn't happen.

#### Summary of changes

Totally remove the dependency on solana-system-interface by inlining
some nonce-related code in solana-message.

* Run cargo-sort

* Remove bincode feature-gating where it's no longer needed

* Fixup imports

* Remove solana_bincode dep, use bincode directly